pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jimeng-image-mcp-server"
version = "1.0.0"
description = "即梦图片生成 MCP 服务器 - 为网站开发提供AI图片生成功能"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
jimeng-image-mcp-server = "jimeng_image_server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jimeng_image_server*"]

[project.urls]
Homepage = "https://github.com/yourusername/jimeng-image-mcp-server"
Repository = "https://github.com/yourusername/jimeng-image-mcp-server"


